ボタンを動的に表示させるための機能の実装途中
