+++ resources/scss/calendar.scss
@@ -1,7 +1,7 @@
 
 
 .month{
-    max-width:1000px;
+    max-width:75%;
     border-top:5px solid gray;
     margin: 0 auto;
 }
@@ -31,6 +31,7 @@ img{
 .month_button{
     display: flex;
     justify-content: space-between;
+    margin: 0 12.5%;
 }
 
 h2{
@@ -62,5 +63,17 @@ h2{
 .schedule_button{
     display: flex;
     justify-content: space-between;
-    margin: 0 200px  50px 200px;
+    margin: 0 20px  20px 20px;
     }
+
+.button{
+    display: flex;
+    flex-wrap: wrap;
+    flex-direction: row;
+    justify-content: center;
+    margin: 0 12.5% 30px 12.5%;
+}
+
+.add_button{
+    margin-left: 0 20px;
+}
